+++ checkpoint/orbax/checkpoint/__init__.py
@@ -72,4 +72,4 @@ except RuntimeError:
 
 # A new PyPI release will be pushed everytime `__version__` is increased.
 # Also modify version and date in CHANGELOG.
-__version__ = '0.5.6'
+__version__ = '0.5.7'